requirements.txt
asgiref==3.5.2
asttokens==2.0.7
attrs==22.1.0
backcall==0.2.0
black==22.6.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1
django-axes==5.40.1
django-environ==0.9.0
django-formtools==2.4
django-ipware==4.0.2
django-otp==1.1.4
django-otp-yubikey==1.0.1
django-phonenumber-field==6.4.0
django-phonenumbers==1.0.1
django-registration==3.3
django-rest-swagger==2.2.0
django-soft-delete==0.9.21
django-softdelete==0.9.4
django-two-factor-auth==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
exceptiongroup==1.1.0
executing==0.9.1
factory-boy==3.2.1
Faker==15.3.4
gerencianet==3.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mercadopago==2.2.0
model-bakery==1.9.0
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==22.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.13.4
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycryptodome==3.16.0
Pygments==2.12.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytest-testdox==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
whitenoise==6.3.0
YubiOTP==1.0.0
